Add Power Tap item, refactor power system to int, geometry ownership + visual fixes
- Add Power Tap (M-sized gear): drone becomes mobile power source
- DroneModel implements IPowerSource; network BFS seeds from drone's room
- PowerNetworkModel: RegisterDrone/UnregisterDrone/NotifyDroneMoved
- Refactor all power types to int (PowerSource, Draw, PowerCost)
- IPowerSource interface extracted from PowerSource class
- Room/passage geometry ownership refactor (per-tile, per-passage half)
- Remove discovered room fog overlay (no gameplay reason)
- Reduce glow intensity (4 -> 1.5), disable GI on emissive materials
- Sync DroneModel.CurrentRoom with DroneController on init and hop
- Fix passage glow tracking for crooked vents

+++ Assets/Scripts/Models/MapModel.cs
@@ -48,7 +48,7 @@ public class MapModel
     public Vector2Int? batteryRoom { get; private set; }
 
     /// <summary>Starting energy for the battery.</summary>
-    public float batteryMaxEnergy { get; private set; } = 100f;
+    public int batteryMaxEnergy { get; private set; } = 100;
 
     /// <summary>A cable edge between two adjacent rooms.</summary>
     public struct CableConnection
@@ -266,7 +266,7 @@ public class MapModel
 
         // Power cable network: battery in hub, cables to station rooms
         batteryRoom = hub;
-        batteryMaxEnergy = 120f;
+        batteryMaxEnergy = 120;
         CableConnections = new List<CableConnection>
         {
             new CableConnection { roomA = hub,  roomB = east },   // powers charging station
